requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain_groq
faiss-cpu
huggingface_hub
sentence_transformers
langchain_huggingface
